Функции,связанные с парсингом и базой данных вынесены в отдельный файл.  Добавлены индексы для полей event и time.
