+++ App/static/txt/test_log
@@ -5049,3 +5049,4 @@ drwxr-xr-x 15 www-data www-data 4096 Ð¼Ð°Ñ   16 16:08 ..
 running pdaemon instances 11504
 15645
 15777
+[2016-05-30 21:47:01.658013559+03:00][9841]:pdaemon is already running.
